+++ src/OpenTelemetry.Exporter.OpenTelemetryProtocol/CHANGELOG.md
@@ -7,6 +7,10 @@ Notes](../../RELEASENOTES.md).
 
 ## Unreleased
 
+## 1.15.2
+
+Released 2026-Apr-08
+
 * Limit how much of the response body is read when export fails and
   error logging is enabled.
   ([#7017](https://github.com/open-telemetry/opentelemetry-dotnet/pull/7017))